+++ pimd/pim_cmd_common.h
@@ -108,6 +108,12 @@ void pim_cmd_show_ip_multicast_helper(struct pim_instance *pim,
 				      struct vty *vty);
 void show_multicast_interfaces(struct pim_instance *pim, struct vty *vty,
 			       json_object *json);
+void show_mroute(struct pim_instance *pim, struct vty *vty,
+			pim_sgaddr *sg, bool fill, bool uj);
+void show_mroute_count(struct pim_instance *pim, struct vty *vty,
+			      bool uj);
+void show_mroute_summary(struct pim_instance *pim, struct vty *vty,
+				json_object *json);
 
 /*
  * Special Macro to allow us to get the correct pim_instance;
